requirements.txt
jicbioimage.core
jicbioimage.transform
jicbioimage.segment
jicbioimage.illustrate
